add if statement for renaming colons
Seems to cause an error in actions all of a sudden. This should hopefully fix it

+++ run.sh
@@ -67,8 +67,13 @@ rm ./manjaro-removed-pkgs.files
 #namely, files with colons would be deleted by repo cleaning process
 #rename command seems to operate differently on Debian
 #don't fully understand, pulled from here https://superuser.com/questions/659876/how-to-rename-files-and-replace-characters
-for f in *:*.pkg.tar.zst; do mv -v "$f" $(echo "$f" | sed 's/:/colon/g'); done
-
+if [ -f ./*:*.pkg.tar.zst ]
+then
+    for f in *:*.pkg.tar.zst; do mv -v "$f" $(echo "$f" | sed 's/:/colon/g'); done
+else
+    :
+fi
+    
 #generate updated releases notes
 cd ../
 sync_date="$(date)"
